refactor(webview): split Inspector label helpers, harden delete test
Extract the DOM-free label markup/validation helpers into
InspectorLabels.ts (Inspector.ts 1005 -> 892 LOC). Fix the inspector
delete test to flush inlineConfirm's deferred outside-click listener
before teardown (was crashing the after-each hook) and no-op the rAF
polyfill.

Fallow: 0 issues / 0 findings / 0 dupes. Remaining informational
split_high_impact target on Inspector.ts is the interaction-controller
density (0.31 vs 0.30) with all three consumers importing the class.
